+++ Assets/Scripts/src/NetStack.Serialization/RawEncoding.cs
@@ -20,7 +20,7 @@ namespace NetStack.Serialization
         where TMemory: struct, IMemory<u32>
     {
         [MethodImpl(Optimization.AggressiveInliningAndOptimization)]
-        public void i32(RawBitWriter<TMemory> b, int value) => b.u32((u32)value, 32);
+        public void i32(RawBitWriter<TMemory> b, i32 value) => b.u32((u32)value, 32);
         
         [MethodImpl(Optimization.AggressiveInliningAndOptimization)]
         public void u32(RawBitWriter<TMemory> b, u32 value) => b.u32(value, 32);
@@ -29,6 +29,6 @@ namespace NetStack.Serialization
         public u32 encode(i32 value) => (u32)value;
 
         [MethodImpl(Optimization.AggressiveInliningAndOptimization)]
-        public void i32(RawBitWriter<TMemory> b, i32 value, i32 numberOfBits) => i32(b, value);
+        public void i32(RawBitWriter<TMemory> b, i32 value, u8 numberOfBits) => i32(b, value);
     }
 }
